backend/requirements.txt
# Core Web Framework
fastapi==0.115.6
uvicorn[standard]==0.34.0
python-multipart==0.0.17

# File Processing
aiofiles==24.1.0
Pillow==11.1.0
# pytesseract removed - Tesseract OCR deprecated
pdf2image==1.17.0
PyPDF2==3.0.1
pdfplumber==0.11.5

# HTTP & Networking
httpx==0.28.1
python-magic==0.4.27

# Data Validation & Models
pydantic==2.10.5
pydantic-settings==2.1.0

# Rate Limiting
slowapi==0.1.9

# AI/ML Integration
openai==1.59.2

# Configuration
python-dotenv==1.0.1

# Database
sqlalchemy==2.0.43
psycopg2-binary==2.9.9
alembic==1.16.5

# Task Queue (for worker health checks)
celery[redis]==5.4.0
redis==5.0.0

# Error Handling & Resilience
tenacity==8.2.3

# Authentication & Security
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0

# NOTE: spaCy moved to worker requirements.txt
# PII removal now happens in worker, not backend

# Testing
pytest==8.3.4
pytest-asyncio==0.25.2
pytest-mock==3.14.0
pytest-cov==6.0.0

# Code Quality & Linting
ruff==0.8.4
mypy==1.13.0
pre-commit==4.0.1
pip-audit==2.7.3

# Type stubs for mypy
types-aiofiles==24.1.0.20240626
types-Pillow==10.2.0.20240822
types-PyYAML==6.0.12.20240917
types-redis==4.6.0.20241004